OTel.API.Context.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

newtype ContextT c m a

Constructors

ContextT

Fields

runContextT :: ContextBackend c -> m a

mapContextT :: forall m n c a b. (m a -> n b) -> ContextT c m a -> ContextT c n b

attachContextValue :: forall m a b. (MonadIO m, MonadMask m) => a -> ContextT a m b -> ContextT a m b

getAttachedContextValue :: forall m a. (MonadIO m, MonadMask m) => ContextT a m (Maybe a)

getAttachedContext :: forall m a. (MonadIO m, MonadThrow m) => ContextT a m Context